feat: shared authenticate response and logout request error handlers
